+++ src/poli/core/abstract_problem_factory.py
@@ -34,7 +34,7 @@ class AbstractProblemFactory(metaclass=MetaProblemFactory):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
         force_isolation: bool = False,
     ) -> Problem:
         """
