+++ src/poli/objective_repository/osimetrinib_mpo/register.py
@@ -65,7 +65,7 @@ class OsimetrinibMPOBlackBox(TDCBlackBox):
 
     Methods
     -------
-    __init__(self, string_representation, force_isolation, batch_size=None, parallelize=False, num_workers=None, evaluation_budget=float("inf")
+    __init__(self, string_representation, force_isolation, batch_size=None, parallelize=False, num_workers=None, evaluation_budget=int("inf")
         Initializes the black box.
 
     References
@@ -87,7 +87,7 @@ class OsimetrinibMPOBlackBox(TDCBlackBox):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
     ):
         super().__init__(
             oracle_name="Osimetrinib_MPO",
@@ -146,7 +146,7 @@ class OsimetrinibMPOProblemFactory(AbstractProblemFactory):
         batch_size: int = None,
         parallelize: bool = False,
         num_workers: int = None,
-        evaluation_budget: int = float("inf"),
+        evaluation_budget: int = int("inf"),
         force_isolation: bool = False,
     ) -> Problem:
         """
